for some reason, the __sync_lock_test_and_set gcc intrinsics is not available for uint8_t, while it's working on uint32_t.

+++ common/rwlock.h
@@ -5,7 +5,7 @@
 #include "util.h"
 
 typedef struct rw_lock_t {
-	uint8_t busy;
+	uint32_t busy;
 	uint8_t writer;
 	uint16_t readercnt;
 } rw_lock;
